backend/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.20.0
python-dotenv>=1.0.0
openai>=0.27.0
aiofiles>=23.1.0
gTTS>=2.3.0